auto-changelog/requirements.txt
typer==0.12.5
rich==13.9.3
GitPython==3.1.43
PyYAML==6.0.2
pytest==8.3.3